refactor(commands): group infra modules into runtime/ to clear DirectorySprawl
